backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
httpx==0.28.1
google-generativeai==0.8.3
google-genai==1.45.0
SQLAlchemy==2.0.35
aiomysql==0.2.0
cryptography==43.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0 
requests==2.31.0
agno==2.1.8
langchain==0.3.7
langchain-community==0.3.5
langchain-text-splitters==0.3.2
chromadb==0.5.20
sentence-transformers==3.3.1
gitpython==3.1.43
python-multipart==0.0.12
tiktoken==0.8.0